Put the OAuth callback journey on the integration boundary
The callback-server suite mixed deterministic URL formatting with real localhost listener and HTTP behavior. Keep the pure URL contract colocated with its source, move only the server journey into the canonical integration tree, retire the resolved semantic-audit exception, and lock in the resulting sanitizer-debt improvement.

Constraint: Integration discovery owns tests/integration, while pure deterministic units remain colocated and the server journey remains Deno-only because it exercises Deno.serve and Deno.listen.

Rejected: Move the whole mixed suite into integration | misclassifies pure URL formatting as an integration concern.

Confidence: high

Scope-risk: narrow

Reversibility: clean

Directive: Keep deterministic callback URL assertions colocated; add real listener, HTTP, origin, and OAuth callback regressions under tests/integration/cli/auth.

Tested: Post-rebase focused split 2 tests / 19 steps; layout 2,396 / 2,345 / 51; semantic audit 2,152 unit / 751 dispositions; sanitizer 388/388. Before the latest unrelated main advance: canonical integration 344 tests / 3,054 steps; test-typecheck 0 new; target format/lint/check; cwd-read, anti-slop, skipped/focused-test ratchets; CLI callback paths passed and three deploy/up failures reproduced identically on clean origin/main.

Not-tested: Hosted CI on the rebased head; live staging is not applicable to a taxonomy-only relocation with no shipped behavior change.
