+++ scripts/lint/check-sanitizer-baseline.ts
@@ -23,10 +23,9 @@ import {
 
 // Lower this when you remove sanitizer opt-outs. Never raise it without a very
 // good reason — a new opt-out means a leak is being suppressed rather than fixed.
-// 396 after moving the dev-server-handlers, api-handler, and RSC
-// hydration/isolation suites onto the in-process request harness, which closes
-// what the old socket servers leaked.
-export const SANITIZER_OPT_OUT_BASELINE = 386;
+// 384 after moving listener-backed suites to canonical integration boundaries
+// and replacing legacy socket servers with in-process request harnesses.
+export const SANITIZER_OPT_OUT_BASELINE = 384;
 
 const OPT_OUT_PATTERN = /sanitize(?:Resources|Ops|Exit)\s*:\s*false/g;
